refactor: move the derivation exports into crate::zcash_jni
Move the nine `RustDerivationTool` exports and `encode_metadata_key`, the
one encoder that serves only them, out of lib.rs. lib.rs drops from 3812 to
3461 lines and from 76 exports to 67.

Another pure move: every function body is relocated unchanged, and the only
lines that are not a straight relocation are the module docstring of the new
file and the import blocks either side of it.

No visibility widening was needed: the helpers these exports call
(`decode_usk`, `parse_network`, `parse_ufvk`, `secret_from_jni`,
`zip32_account_index_from_jlong`) live in the crate root, and private items
there are already in scope for a descendant module.

Nine imports that only the derivation code used are dropped from lib.rs:
the whole `zcash_address::unified` group, `NetworkConstants`, and four
`zip32` items.

Exported symbol set unchanged: 78 before, 78 after, empty `nm -gU` diff.

    ./scripts/check-jni-symbols.sh          # 78 symbols, baseline matches
    ./scripts/check-jni-move.sh d65b4d2c    # 78 byte-identical, 0 changed
